Renamed EnrollmentModal => EnrollmentDialog
